feat(DTFS2-7795): move macro component tests into correct directory
